Remove data and lang
Rather than perform storage for sync (the data parameter), use indexeddb or cache.  This also encourages developers to use a single event for multiple operations (e.g., store lots of tweets in IndexedDB and send them all when next online in a single sync).

Remove lang because that can be determined from the page.