Rewrite spec using bikeshed.
Besides changing the tool used, this also rephrases some of the algorithms,
and tries to make some behavior more explicit.

spec/index.html is autogenerated from spec/index.bs using the bikeshed tool.

The new footer.include and default.css files should be temporary until we've
figure out what the corect template should be for whatever group this spec is
going to be developed in.