Change the check in register to reject if no top-level clients exist.
There is no need to special case the type of the context this method was called
in; calling register from a top-level window means that at least one top-level
client must exist, so this is automatically allowed by just checking the frame
types of all clients at the origin. This addresses issue #114.