Update the definition of a sync event running "in the background".
This now excludes the case when the event is running and there's at
least one nested frame for the origin. A nested frame will have a window
client and thus running a sync event for these won't cause the privacy
concerns this definition is used to describe.

This change also aligns this spec with the Periodic Background Sync
prose.